Minor Bug Fixes and cp in complete, cp out started
Minor Bug Fixes including a cout debug flag messing up the ls function. The cp out implementation works with all VDIs given, dynamic and static, 1k and 4k page sizes. Started implementation for cp in, to copy a file into the VDI from the host. (The cp function is in reference to the VDI.)