Continued work on cp in functionality
Added functions, constants, and prototypes for more helper functions and towards the functionality of cp on; copying into the VDI. Checking the max size of the file. Functions to read and write bitmaps, create a directory, Started with writing file function. Heavily commented. Have placeholders in a few areas to be removed and filled in correctly later. Not at a state where successfully compiling; ironing out errors in the AM.